Relocate code to new file
