+++ pack3d/model.go
@@ -7,14 +7,6 @@ import (
 	"github.com/fogleman/fauxgl"
 )
 
-// Restricted mode means:
-//
-// 1. Assume there is a 3D print floor.
-// 2. The Z axis is upward.
-// 3. The only transformations allowed are moving along X and Y and rotating around Z axis.
-// 4. The bottom of all 3D models are aligned on the 3D print floor.
-var Restricted bool = true
-
 var Rotations []fauxgl.Matrix
 
 func init() {
@@ -232,20 +224,4 @@ func (m *Model) Copy() Annealable {
 		items[i] = item.Copy()
 	}
 	return &Model{items, m.MinVolume, m.MaxVolume, m.Deviation}
-}
-
-// Random unit vector inside XY plane.
-// Its Z component is zero.
-// Reference:
-// https://github.com/fogleman/fauxgl/blob/27cddc103802008bbda73dc74cab49038d96fdf3/vector.go#L16
-func RandomUnitVecXY() fauxgl.Vector {
-	for {
-		x := rand.Float64()*2 - 1
-		y := rand.Float64()*2 - 1
-		z := float64(0)
-		if x*x+y*y+z*z > 1 {
-			continue
-		}
-		return fauxgl.Vector{X: x, Y: y, Z: z}.Normalize()
-	}
 }
